requirements.txt
requests>=2.31.0
feedparser>=6.0.10
google-generativeai>=0.3.0
